tools/requirements.txt
Pillow>=9.0.0
rembg>=2.0.50
